+++ textract/parsers/odt_parser.py
@@ -23,9 +23,8 @@ class Parser(BaseParser):
     def to_string(self):
         """ Converts the document to a string. """
         buff = u""
-        for val in ["text:p", "text:h", "text:list"]:
-            for paragraph in self.content.getElementsByTagName(val):
-                buff += self.text_to_string(paragraph) + "\n"
+        for paragraph in self.content.getElementsByTagName("text:p"):
+            buff += self.text_to_string(paragraph) + "\n"
         return buff
 
     def text_to_string(self, element):